Extract FavoritesAndPronunciationsRepository from VocabularyRepository
