+++ app/src/main/kotlin/de/ywegel/svenska/data/FavoritesAndPronunciationsRepository.kt
@@ -1,0 +1,9 @@
+package de.ywegel.svenska.data
+
+import de.ywegel.svenska.data.model.Vocabulary
+
+interface FavoritesAndPronunciationsRepository {
+    fun getFavorites(containerId: Int?): List<Vocabulary>
+
+    fun getPronunciations(containerId: Int?): List<Vocabulary>
+}
