+++ app/src/test/kotlin/de/ywegel/svenska/fakes/VocabularyRepositoryFake.kt
@@ -1,6 +1,7 @@
 package de.ywegel.svenska.fakes
 
 import de.ywegel.svenska.data.ContainerRepository
+import de.ywegel.svenska.data.FavoritesAndPronunciationsRepository
 import de.ywegel.svenska.data.SortOrder
 import de.ywegel.svenska.data.VocabularyRepository
 import de.ywegel.svenska.data.model.Vocabulary
@@ -13,7 +14,7 @@ import kotlinx.coroutines.runBlocking
 class VocabularyRepositoryFake(
     initialVocabulary: List<Vocabulary> = emptyList(),
     initialContainers: List<VocabularyContainer> = emptyList(),
-) : VocabularyRepository, ContainerRepository {
+) : VocabularyRepository, ContainerRepository, FavoritesAndPronunciationsRepository {
 
     constructor(
         initialContainers: List<VocabularyContainer> = emptyList(),
